know memoization and use React.memo()
